added the new bookmarkicon
