+++ pages/bestellung/ruecktritt.js
@@ -1,21 +1,22 @@
 import styled from "styled-components";
 import Link from "next/link";
-import StyledIcon from "../../components/BookmarkIcon";
 import {CopyToClipboard} from "react-copy-to-clipboard";
 import Breadcrumb from "../../components/Breadcrumb";
 import ReturnButton from "../../components/ReturnButton";
 import PageContainer from "../../components/PageContainer";
 import CopyButton from "../../components/CopyButton";
 import StyledArticle from "../../components/Article";
+import BookmarkIcon from "../../components/BookmarkIcon";
 
 export default function RuecktrittPage() {
   const path = "/online bestellen und kaufen/Rücktritt";
   return (
+    <>
+    <BookmarkIcon />
     <PageContainer>
       <StyledLink href="/bestellung">
         <Breadcrumb path={path} />
       </StyledLink>
-      <StyledIcon icon="mingcute:bookmark-fill" color="#572887" />
       <StyledArticle>
         <p>
           {`Ich möchte vom Kaufvertrag zurück treten.
@@ -46,6 +47,7 @@ Für weitere Infos dazu, klicke auf die jeweiligen Links weiter unten.
         <ReturnButton />
       </StyledLink>
     </PageContainer>
+    </>
   );
 }
